+++ transcriptic/__init__.py
@@ -1,32 +1,29 @@
-from transcriptic.objects import Run, Project, Aliquot, Resource
+from __future__ import print_function
+from transcriptic.objects import Run, Project, Aliquot, Resource, _check_ctx
 from transcriptic.objects import Container, Dataset, ProtocolPreview
+from transcriptic.config import Connection
 
 ctx = None
 
 
-def _get_object(obj_id, klass):
-    data = ctx._get_object(obj_id)
-    return klass(data['id'], data, connection=ctx)
-
-
 def run(obj_id):
-    return _get_object(obj_id, Run)
+    return Run(obj_id)
 
 
 def project(obj_id):
-    return _get_object(obj_id, Project)
+    return Project(obj_id)
 
 
 def resource(obj_id):
-    return _get_object(obj_id, Resource)
+    return Resource(obj_id)
 
 
 def aliquot(obj_id):
-    return _get_object(obj_id, Aliquot)
+    return Aliquot(obj_id)
 
 
 def container(obj_id):
-    return _get_object(obj_id, Container)
+    return Container(obj_id)
 
 
 def preview(protocol):
@@ -41,5 +38,13 @@ def submit(protocol, project_id, title=None, test_mode=False):
     return ctx.submit_run(protocol, project_id=project_id, title=title, test_mode=test_mode)
 
 
-def dataset(obj_id, key="*"):
-    return ctx.dataset(obj_id=obj_id, key=key)
+def dataset(data_id, key="*"):
+    return ctx.dataset(data_id=data_id, key=key)
+
+
+def connect(transcriptic_path="~/.transcriptic"):
+    #TODO: Mirror login code from CLI
+    try:
+        ctx = Connection.from_file(transcriptic_path)
+    except:
+        print ("Unable to find .transcriptic file, please ensure the right path is provided")