+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## `transcriptic` Changelog
 
 ## Unreleased
+Changed
+- Objects module has been heavily reworked and documentation added. This is especially true for Project, Run and Dataset objects
+
 Fixed
 - Fixed package related CLI issues